2004-06-25  Stuart Hastings <stuart@apple.com>
	Radar 3680191
	* gcc/tree-nested.c: Cast STACK_BOUNDARY for portability.
	* gcc/config/i386/i386.c (output_pic_addr_const): Stubify optimized symbols.
	* gcc/target-def.h: (TARGET_BECT_BUILD_VECTOR_INIT) Initialize correctly.


git-svn-id: svn+ssh://gcc.gnu.org/svn/gcc/branches/apple-ppc-branch@83691 138bc75d-0d04-0410-961f-82ee72b054a4

+++ gcc/tree-nested.c
@@ -396,7 +396,8 @@ get_trampoline_type (void)
 
   /* If we won't be able to guarantee alignment simply via TYPE_ALIGN,
      then allocate extra space so that we can do dynamic alignment.  */
-  if (align > STACK_BOUNDARY)
+  /* APPLE LOCAL STACK_BOUNDARY must be a signed expression on Darwin/x86 */
+  if (align > (unsigned int) STACK_BOUNDARY)
     {
       size += ((align/BITS_PER_UNIT) - 1) & -(STACK_BOUNDARY/BITS_PER_UNIT);
       align = STACK_BOUNDARY;
